requirements.txt
rel~=0.4.7
websocket-client~=1.4.1
pynput~=1.7.6